Move in-memory storage implementation to helpers
